Changed from uint16 to uint64 for maximum recursion depth
Also now using single-precision floating point numbers (float) for data file

+++ MandelbrotSet/DistributedComputation.cs
@@ -17,7 +17,7 @@ namespace MandelbrotSet
             double minI,
             double maxI,
             double stepI,
-            ushort maximumRecursionDepth,
+            ulong maximumRecursionDepth,
             ulong bufferSize)
         {
             
@@ -31,7 +31,7 @@ namespace MandelbrotSet
             double minI,
             double maxI,
             double stepI,
-            ushort maximumRecursionDepth,
+            ulong maximumRecursionDepth,
             ulong bufferSize)
         {
